% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{as.LAS}
\alias{as.LAS}
\title{Convert data.frame of points to LAS object of lidR package.}
\usage{
as.LAS(df, crs = sf::NA_crs_)
}
\description{
Convert data.frame of points (received from PointDB or PointCloud) to \link[lidR]{LAS} in lidR package.
}
\details{
Optional parameter crs: crs of points. e.g. crs <- sf::st_crs(paste0('EPSG:', pointdb$info$epsg))
}
\seealso{
\link[lidR]{LAS}
}
\author{
woellauer
}
