% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointDB_functions.R
\name{clean_data}
\alias{clean_data}
\title{Tries to remove erroneous points.}
\usage{
clean_data(df)
}
\arguments{
\item{df}{data.frame of points}
}
\value{
cleaned data.frame
}
\description{
The lowest and highest z-value points will be removed (1\% each). So most outliers are not contained in the resulting data.frame.
}
\author{
woellauer
}
