% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{extent_diameter}
\alias{extent_diameter}
\title{create extent by position and edge length.}
\usage{
extent_diameter(x, y, d)
}
\arguments{
\item{x}{first coordinate}

\item{y}{second coordinate}

\item{d}{diameter (edge length)}
}
\value{
extent
}
\description{
Creates an \link{extent} that covers the square with center point x,y and with diameter (edge length) d.
}
\details{
Extents are used to query data from \link{RasterDB} and from \link{PointDB}.
}
\seealso{
\link{extent}
}
\author{
woellauer
}
