% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{extent_radius}
\alias{extent_radius}
\title{create extent by position and radius}
\usage{
extent_radius(x, y, r)
}
\arguments{
\item{x}{first coordinate}

\item{y}{second coordinate}

\item{r}{radius}
}
\value{
extent
}
\description{
Creates an \link{extent} that covers the square over the circle at point x,y with radius r.
}
\details{
Extents are used to query data from \link{RasterDB} , from \link{PointCloud} and from \link{PointDB}.
}
\seealso{
\link{extent}
}
\author{
woellauer
}
