% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoteSensing.R
\docType{class}
\name{RemoteSensing}
\alias{RemoteSensing}
\title{RemoteSensing class}
\format{
RemoteSensing \code{\link{R6Class}} object.

Instance objects of R6Class are created by 'new':

\code{remotesensing <- RemoteSensing$new(url, userpwd=NULL)}

Methods of instance objects are called by '$':

\code{result <- remotesensing$method(parameters)}
}
\usage{
remotesensing <- RemoteSensing$new(url, userpwd=NULL)

remotesensing$rasterdbs
remotesensing$rasterdb(name)
remotesensing$create_rasterdb(name, proj4=NULL, resolution=NULL, storage_type=NULL, create_new=TRUE)

remotesensing$pointclouds
remotesensing$pointcloud(name)

remotesensing$pointdbs
remotesensing$pointdb(name)

remotesensing$vectordbs
remotesensing$vectordb(name)

remotesensing$roi_groups
remotesensing$roi_group(name)
remotesensing$roi(group_name, roi_name)

remotesensing$poi_groups
remotesensing$poi_group(name)
remotesensing$poi(group_name, poi_name)

remotesensing$tasks
remotesensing$tasks_rasterdb
remotesensing$tasks_pointdb
remotesensing$tasks_pointcloud
remotesensing$tasks_voxeldb
remotesensing$tasks_vectordb
remotesensing$submit_task(task, wait = TRUE, polling_interval = 0.5)

remotesensing$web()
}
\description{
Remote sensing database manages (hyperspectral) \strong{rasters}, (LiDAR) \strong{point-clouds}, \strong{voxel} data and \strong{vector} data
as well as auxiliary ROIs (regions of interest as named polygons) and POIs (points of interest as named points).
}
\details{
Objects of RemoteSensing class encapsulate connections to one remote sensing database.
}
\section{Methods}{


\describe{

\item{RemoteSensing$new(url, userpwd=NULL)}{Open Remote Sensing Database.

url: url of server. (local e.g. "http://localhost:8081" or remote e.g. "http://example.com:8081")

userpwd: optional authentication (format: "user:password")

returns: RemoteSensing object}

\item{$rasterdbs}{get names of RasterDBs contained in Remote Sensing Database.}

\item{$rasterdb(name)}{get RasterDB by name.}

\item{$create_rasterdb(name, proj4=NULL, resolution=NULL, storage_type=NULL, create_new=TRUE)}{creates new empty RasterDB.

name: name of RasterDB layer

proj4: (optional) CRS as PROJ4

resolution: (optional) one number or vector of two numbers of pixel size in projection units, e.g. 0.5 -> xres=0.5, yres=0.5  or e.g. c(10, 5) -> xres=10, yres=5

storage_type: (optional) storage_type RasterUnit (default) or TileStorage

create_new: (optional) TRUE: delete old raster layer of same name, FALSE: if raster layer already exists do not create it

returns: success message}

\item{$pointclouds}{get names of PointClouds contained in Remote Sensing Database.}

\item{$pointcloud(name)}{get PointCloud by name.}

\item{$pointdbs}{get names of PointDBs contained in Remote Sensing Database.}

\item{$pointdb(name)}{get PointDB by name.}

\item{$vecordbs}{get names of VectorDBs contained in Remote Sensing Database.}

\item{$vectordb(name)}{get VectorDB by name.}

\item{$roi_groups}{get names of ROI-groups.

returns: data.frame of name and description}

\item{$roi_group(name)}{get ROI-group by name.

returns: data.frame of ROI-name and polygon}

\item{$roi(group_name, roi_name)}{get one ROI.

returns: polygon}

\item{$poi_groups}{get names of POI-groups.

returns: data.frame of name and description}

\item{$poi_group(name)}{get POI-group by name.

returns: data.frame of POI-name and position}

\item{$poi(group_name, poi_name)}{get one POI.

returns: position}

\item{$tasks}{Get details of tasks that can be executed by 'submit_task'.

returns: data.frame with nested data.frames of task details}

\item{$tasks_rasterb}{Get details of type 'task_rasterb' tasks that can be executed by 'submit_task'.

returns: data.frame with task details}

\item{$tasks_pointdb}{Get details of type 'task_pointdb' tasks that can be executed by 'submit_task'.

returns: data.frame with task details}

\item{$tasks_pointcloud}{Get details of type 'task_pointcloud' tasks that can be executed by 'submit_task'.

returns: data.frame with task details}

\item{$tasks_voxeldb}{Get details of type 'task_voxeldb' tasks that can be executed by 'submit_task'.

returns: data.frame with task details}

\item{$tasks_vectordb}{Get details of type 'task_vectordb' tasks that can be executed by 'submit_task'.

returns: data.frame with task details}

\item{$submit_task(task, wait = TRUE, polling_interval = 0.5)}{Submit a remote task to RSDB server.

task: remote task parameters. List of parameters or character containing parameters as JSON text.

wait: wait until task finished. Then return result. Default: TRUE.

polling_interval: if wait = TRUE, check task status every polling_interval seconds.

returns: RemoteTask object

If wait = FALSE, on the returned RemoteTask object $wait() needs to be called to check for task finish. If wait = TRUE that wait() is called internally.}

\item{$web()}{Open web interface in browser.}
}
}

\section{ROI}{

Region of interest (ROI) is a named polygon.

Matrix of one row per polygon vertex and two columns with x- and y-coordinates describe the polygon.

First and last vertex need to be same (a closed polygon).

example:
\preformatted{
# get one ROI
roi <- remotesensing$roi(group_name="kili_A", roi_name="cof3_A")

# get name of ROI
roi$name

# get matrix of points
roi$polygon[[1]]
}
}

\section{POI}{

Point of interest (POI) is a named point.

example:
\preformatted{
# get one POI
poi <- remotesensing$poi(group_name="kili", poi_name="cof3")

# get name of ROI
poi$name

# get x-coodinate of POI
poi$x

# get y-coodinate of POI
poi$y
}
}

\examples{
# open remote sensing database
library(RSDB)
# remotesensing <- RemoteSensing$new("http://localhost:8081", "user:password") # local
remotesensing <- RemoteSensing$new("http://example.com:8081", "user:password") # remote server

# get names of RasterDBs
remotesensing$rasterdbs

# get one rasterdb
rasterdb <- remotesensing$rasterdb("kili_campaign1")

# get names of PointDBs
remotesensing$pointdbs

# get one pointdb
pointdb <- remotesensing$pointdb("kili")

# get names of ROI groups
remotesensing$roi_groups

# get one ROI group
rois <- remotesensing$roi_group("kili_A")

# get one ROI
roi <- remotesensing$roi(group_name="kili_A", roi_name="cof3_A")

# get names of POI groups
remotesensing$poi_groups

# get one POI group
pois <- remotesensing$poi_group("kili")

# get one POI
poi <- remotesensing$poi(group_name="kili", poi_name="cof3")

# create extent around POI of 10 meter edge length
ext <- extent_diameter(poi$x, poi$y, 10)

# get RasterStack of all bands at ext
r <- rasterdb$raster(ext)

# get data.frame of LiDAR points at ext
df1 <- pointdb$query(ext)

#get data.frame of LiDAR points at polygon ROI
df2 <- pointdb$query_polygon(roi$polygon[[1]])

# open web interface in browser
remotesensing$web()

}
\seealso{
\link{RasterDB} \link{PointCloud} \link{PointDB} \link{VectorDB} \link{VoxelDB} \link{RemoteTask}
}
\author{
woellauer
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RemoteSensing-new}{\code{RemoteSensing$new()}}
\item \href{#method-RemoteSensing-lidar}{\code{RemoteSensing$lidar()}}
\item \href{#method-RemoteSensing-pointdb}{\code{RemoteSensing$pointdb()}}
\item \href{#method-RemoteSensing-rasterdb}{\code{RemoteSensing$rasterdb()}}
\item \href{#method-RemoteSensing-lidar_layers}{\code{RemoteSensing$lidar_layers()}}
\item \href{#method-RemoteSensing-web}{\code{RemoteSensing$web()}}
\item \href{#method-RemoteSensing-poi_group}{\code{RemoteSensing$poi_group()}}
\item \href{#method-RemoteSensing-poi}{\code{RemoteSensing$poi()}}
\item \href{#method-RemoteSensing-roi_group}{\code{RemoteSensing$roi_group()}}
\item \href{#method-RemoteSensing-roi}{\code{RemoteSensing$roi()}}
\item \href{#method-RemoteSensing-create_rasterdb}{\code{RemoteSensing$create_rasterdb()}}
\item \href{#method-RemoteSensing-pointcloud}{\code{RemoteSensing$pointcloud()}}
\item \href{#method-RemoteSensing-vectordb}{\code{RemoteSensing$vectordb()}}
\item \href{#method-RemoteSensing-voxeldb}{\code{RemoteSensing$voxeldb()}}
\item \href{#method-RemoteSensing-submit_task}{\code{RemoteSensing$submit_task()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RemoteSensing-new"></a>}}
\if{latex}{\out{\hypertarget{method-RemoteSensing-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RemoteSensing$new(url, userpwd = NULL, ssl_verifypeer = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RemoteSensing-lidar"></a>}}
\if{latex}{\out{\hypertarget{method-RemoteSensing-lidar}{}}}
\subsection{Method \code{lidar()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RemoteSensing$lidar(layer)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RemoteSensing-pointdb"></a>}}
\if{latex}{\out{\hypertarget{method-RemoteSensing-pointdb}{}}}
\subsection{Method \code{pointdb()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RemoteSensing$pointdb(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RemoteSensing-rasterdb"></a>}}
\if{latex}{\out{\hypertarget{method-RemoteSensing-rasterdb}{}}}
\subsection{Method \code{rasterdb()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RemoteSensing$rasterdb(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RemoteSensing-lidar_layers"></a>}}
\if{latex}{\out{\hypertarget{method-RemoteSensing-lidar_layers}{}}}
\subsection{Method \code{lidar_layers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RemoteSensing$lidar_layers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RemoteSensing-web"></a>}}
\if{latex}{\out{\hypertarget{method-RemoteSensing-web}{}}}
\subsection{Method \code{web()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RemoteSensing$web()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RemoteSensing-poi_group"></a>}}
\if{latex}{\out{\hypertarget{method-RemoteSensing-poi_group}{}}}
\subsection{Method \code{poi_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RemoteSensing$poi_group(group_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RemoteSensing-poi"></a>}}
\if{latex}{\out{\hypertarget{method-RemoteSensing-poi}{}}}
\subsection{Method \code{poi()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RemoteSensing$poi(group_name, poi_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RemoteSensing-roi_group"></a>}}
\if{latex}{\out{\hypertarget{method-RemoteSensing-roi_group}{}}}
\subsection{Method \code{roi_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RemoteSensing$roi_group(group_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RemoteSensing-roi"></a>}}
\if{latex}{\out{\hypertarget{method-RemoteSensing-roi}{}}}
\subsection{Method \code{roi()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RemoteSensing$roi(group_name, roi_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RemoteSensing-create_rasterdb"></a>}}
\if{latex}{\out{\hypertarget{method-RemoteSensing-create_rasterdb}{}}}
\subsection{Method \code{create_rasterdb()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RemoteSensing$create_rasterdb(
  name,
  proj4 = NULL,
  resolution = NULL,
  storage_type = NULL,
  create_new = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RemoteSensing-pointcloud"></a>}}
\if{latex}{\out{\hypertarget{method-RemoteSensing-pointcloud}{}}}
\subsection{Method \code{pointcloud()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RemoteSensing$pointcloud(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RemoteSensing-vectordb"></a>}}
\if{latex}{\out{\hypertarget{method-RemoteSensing-vectordb}{}}}
\subsection{Method \code{vectordb()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RemoteSensing$vectordb(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RemoteSensing-voxeldb"></a>}}
\if{latex}{\out{\hypertarget{method-RemoteSensing-voxeldb}{}}}
\subsection{Method \code{voxeldb()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RemoteSensing$voxeldb(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RemoteSensing-submit_task"></a>}}
\if{latex}{\out{\hypertarget{method-RemoteSensing-submit_task}{}}}
\subsection{Method \code{submit_task()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RemoteSensing$submit_task(task, wait = TRUE, polling_interval = 0.5)}\if{html}{\out{</div>}}
}

}
}
