% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{roi_to_extent}
\alias{roi_to_extent}
\title{Convert ROI polygon to rectangular extent}
\usage{
roi_to_extent(roi)
}
\arguments{
\item{roi}{a roi returned by database}
}
\value{
extent
}
\description{
Convert ROI polygon to rectangular extent
}
\seealso{
\link{extent}
}
\author{
woellauer
}
