% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VectorDB.R
\name{VectorDB}
\alias{VectorDB}
\title{VectorDB class}
\format{VectorDB \link{R6Class} object.}
\usage{
# connect to remote sensing database
remotesensing <- RemoteSensing$new(url, userpwd=NULL)

# get VectorDB names
remotesensing$vectordbs

# open one VectorDB in connected remote sensing database
vectordb <- remotesensing$vectordb(name)

# operate on opened VectorDB:

vectordb$getVectors(epsg=NULL)
vectordb$getVectorsWGS84()

vectordb$attributes
vectordb$description
vectordb$epsg
vectordb$meta
vectordb$name
vectordb$name_attribute
vectordb$proj4
vectordb$title
}
\description{
VectorDB class provides methods to access to vector data of Remote Sensing Database.
}
\section{Details}{


In the Following methods of VectorDB are described. Use instance objects to call that methods (e.g. vecotdb$method(paramters) ).
Use instance of \link{RemoteSensing} class to create instances of VectorDB.

\describe{

\item{$getVectors(epsg=NULL)}{Get vector data.

returns: vector data

parameters:

\strong{epsg}: desired reprojection.

EPSG code of desired reprojection. Code needs to be number as character. e.g. '4326'. If paramter is missing data is returned in original projection.


}

\item{$getVectorsWGS84()}{Get vector data in WGS84 longlat projection (EPSG:4326). Shortcut of vectordb$getVectors('4326').

returns: vector data

}

}
}

\examples{

library(RSDB)
#remotesensing <- RemoteSensing$new("http://localhost:8081", "user:password") # local
remotesensing <- RemoteSensing$new("http://example.com:8081", "user:password") # remote server

# get VectorDB names
vectordbs <- remotesensing$vectordbs

# open vectordb
vectordb <- remotesensing$vectordb("ep_plot_polygons")

# get vecotor data in WGS84 longlat projection (EPSG:4326)
data <- vectordb$getVectorsWGS84()

# static visualize vector data
plot(data)

# dynamic visualize vector data
library(mapview)
mapview::mapview(data)

}
\seealso{
\link{RemoteSensing}
}
\author{
woellauer
}
\keyword{line}
\keyword{point}
\keyword{polygon}
\keyword{vector}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VectorDB$new()}}
\item \href{#method-getVectors}{\code{VectorDB$getVectors()}}
\item \href{#method-getVectorsWGS84}{\code{VectorDB$getVectorsWGS84()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorDB$new(base_url, name, curlHandle, rsdbConnector)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getVectors"></a>}}
\subsection{Method \code{getVectors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorDB$getVectors(epsg = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getVectorsWGS84"></a>}}
\subsection{Method \code{getVectorsWGS84()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorDB$getVectorsWGS84()}\if{html}{\out{</div>}}
}

}
}
