% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{visualise_raster}
\alias{visualise_raster}
\title{Interactively visualise RasterLayer in 3d.}
\usage{
visualise_raster(rasterLayer)
}
\arguments{
\item{rasterLayer}{RasterLayer received by e.g. pointdb$query_raster(...)}
}
\description{
Visualises a RasterLayer (Matrix) as interactive 3d surface with cell values as elevation.
}
\author{
woellauer
}
