% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VoxelDB.R
\name{VoxelDB}
\alias{VoxelDB}
\title{R6 Class Representing a VoxelDB layer}
\description{
A VoxelDB layer contains voxel data.
}
\details{
Instance object of VoxelDB class are created by instances of RemoteSensing class.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VoxelDB-new}{\code{VoxelDB$new()}}
\item \href{#method-VoxelDB-voxels}{\code{VoxelDB$voxels()}}
\item \href{#method-VoxelDB-aggregated_voxels}{\code{VoxelDB$aggregated_voxels()}}
\item \href{#method-VoxelDB-raster}{\code{VoxelDB$raster()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VoxelDB-new"></a>}}
\if{latex}{\out{\hypertarget{method-VoxelDB-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VoxelDB$new(rsdbConnector, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VoxelDB-voxels"></a>}}
\if{latex}{\out{\hypertarget{method-VoxelDB-voxels}{}}}
\subsection{Method \code{voxels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VoxelDB$voxels(
  ext = NULL,
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL,
  zmin = NULL,
  zmax = NULL,
  product = "count",
  time_slice_id = NULL,
  time_slice_name = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VoxelDB-aggregated_voxels"></a>}}
\if{latex}{\out{\hypertarget{method-VoxelDB-aggregated_voxels}{}}}
\subsection{Method \code{aggregated_voxels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VoxelDB$aggregated_voxels(
  ext = NULL,
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL,
  zmin = NULL,
  zmax = NULL,
  time_slice_id = NULL,
  time_slice_name = NULL,
  aggregation_factor = NULL,
  aggregation_factor_x = NULL,
  aggregation_factor_y = NULL,
  aggregation_factor_z = NULL,
  crop = FALSE,
  product = "sum"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VoxelDB-raster"></a>}}
\if{latex}{\out{\hypertarget{method-VoxelDB-raster}{}}}
\subsection{Method \code{raster()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VoxelDB$raster(
  ext,
  zmin = NULL,
  zmax = NULL,
  time_slice_id = NULL,
  time_slice_name = NULL,
  aggregation_factor = NULL,
  aggregation_factor_x = NULL,
  aggregation_factor_y = NULL,
  product = "sum"
)}\if{html}{\out{</div>}}
}

}
}
