% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostGIS.R
\name{PostGIS}
\alias{PostGIS}
\title{PostGIS class}
\format{
PostGIS \link{R6Class} object.
}
\usage{
# connect to remote sensing database
remotesensing <- RemoteSensing$new(url, userpwd=NULL)

# open one PostGIS table in connected remote sensing database
postgis <- remotesensing$postgis(name)

# operate on opened postgis:
}
\description{
PostGIS class provides methods to access PostGIS tables of Remote Sensing Database.

Use instance of \link{RemoteSensing} class to create instances of PostGIS class.

In section 'Active bindings' class properties and in section 'Methods' class functions of PostGIS class are described.

Use instance objects of PostGIS class to call that functionality (e.g. postgis$name ).
}
\examples{

# open remote sensing database
library(RSDB)
#remotesensing <- RemoteSensing$new("http://localhost:8081", "user:password") # local
remotesensing <- RemoteSensing$new("http://example.com:8081", "user:password") # remote server


}
\seealso{
\link{RemoteSensing}
}
\author{
woellauer
}
\keyword{database}
\keyword{remote-sensing}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{PostGIS layer name.}

\item{\code{meta}}{Meta data.}

\item{\code{epsg}}{EPSG code.}

\item{\code{fields}}{database table fields.}

\item{\code{class_fields}}{database table fields that are marked as class fields.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PostGIS-new}{\code{PostGIS$new()}}
\item \href{#method-PostGIS-getFeatures}{\code{PostGIS$getFeatures()}}
\item \href{#method-PostGIS-getRaster}{\code{PostGIS$getRaster()}}
\item \href{#method-PostGIS-getIndices}{\code{PostGIS$getIndices()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PostGIS-new"></a>}}
\if{latex}{\out{\hypertarget{method-PostGIS-new}{}}}
\subsection{Method \code{new()}}{
Use instance of \link{RemoteSensing} class to create instances of PostGIS class (corresponding to one table in a PostGIS database).
This initialize function is for internal use only.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PostGIS$new(rsdbConnector, name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rsdbConnector}}{rsdb server connection.}

\item{\code{name}}{PostGIS table name.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PostGIS-getFeatures"></a>}}
\if{latex}{\out{\hypertarget{method-PostGIS-getFeatures}{}}}
\subsection{Method \code{getFeatures()}}{
Get vector features and properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PostGIS$getFeatures(bbox = NULL, crop = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bbox}}{Bounding box.}

\item{\code{crop}}{Crop vector features to bbox.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PostGIS-getRaster"></a>}}
\if{latex}{\out{\hypertarget{method-PostGIS-getRaster}{}}}
\subsection{Method \code{getRaster()}}{
Get raster from rasterized vector features.

Raster pixel values are the values of field parameter.

If priority parameter = NULL, order of feature painting is unspecified, so for overlapping features of different field values the final pixel value is the value of one of the overlapping features.

The priority parameter specifies order of feature painting. For overlapping features the last entry in the priority vector determines the pixel value.
Only features values listed in the priority vector are painted, all other features are omitted in the final raster.

Example:

feature field values: 1 2 3 4

priority = c(2,3)  => Only features of value 2 and 3 are painted. For overlapping features value 3 is the pixel value.

priority = c(3,2)  => Only features of value 2 and 3 are painted. For overlapping features value 2 is the pixel value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PostGIS$getRaster(bbox, field = NULL, dx = 1, dy = 1, priority = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bbox}}{Bounding box.}

\item{\code{field}}{Value field for pixel values.}

\item{\code{dx}}{Pixel x resolution.}

\item{\code{dy}}{Pixel y resolution.}

\item{\code{priority}}{Features rasterization order.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PostGIS-getIndices"></a>}}
\if{latex}{\out{\hypertarget{method-PostGIS-getIndices}{}}}
\subsection{Method \code{getIndices()}}{
Calculate indices from vector features.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PostGIS$getIndices(bbox, field = NULL, crop = TRUE, values = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bbox}}{Bounding box.}

\item{\code{field}}{Classification field for feature aggregation.}

\item{\code{crop}}{Crop vector features to bbox.}

\item{\code{values}}{If raw values should be returned instead of summary statistical measures per class.}
}
\if{html}{\out{</div>}}
}
}
}
