% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoteTask.R
\name{RemoteTask}
\alias{RemoteTask}
\title{RemoteTask class}
\format{
RemoteTask \link{R6Class} object.
}
\usage{
# Connect to remote sensing database.
remotesensing <- RemoteSensing$new(url, userpwd=NULL)

# Run a remote task at RSDB server.
remote_task <- remotesensing$submit_task(task)

# Get current task status information.
status <- remote_task$status

# Wait until task has finished processing.
status <- remote_task$wait()
}
\description{
RemoteTask class provides methods to manage a remote task running at RSDB server.
}
\examples{

remotesensing <- RSDB::RemoteSensing$new('http://localhost:8080', userpwd='user:password')

task <- list(task_rasterdb = 'count_pixels', rasterdb = 'raster1')

remote_task <- remotesensing$submit_task(task)

status <- remote_task$wait()

status$message

}
\seealso{
\link{RemoteSensing}
}
\author{
woellauer
}
\keyword{remote}
\keyword{task}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RemoteTask$new()}}
\item \href{#method-wait}{\code{RemoteTask$wait()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RemoteTask$new(rsdbConnector, id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wait"></a>}}
\if{latex}{\out{\hypertarget{method-wait}{}}}
\subsection{Method \code{wait()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RemoteTask$wait()}\if{html}{\out{</div>}}
}

}
}
