% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VoxelDB.R
\name{VoxelDB}
\alias{VoxelDB}
\title{R6 Class Representing a VoxelDB layer}
\description{
A VoxelDB layer contains voxel data.
}
\details{
Instance object of VoxelDB class are created by instances of RemoteSensing class.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VoxelDB$new()}}
\item \href{#method-voxels}{\code{VoxelDB$voxels()}}
\item \href{#method-aggregated_voxels}{\code{VoxelDB$aggregated_voxels()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VoxelDB$new(rsdbConnector, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-voxels"></a>}}
\if{latex}{\out{\hypertarget{method-voxels}{}}}
\subsection{Method \code{voxels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VoxelDB$voxels(
  ext = NULL,
  x = NULL,
  y = NULL,
  z = NULL,
  product,
  time_slice_id = NULL,
  time_slice_name = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aggregated_voxels"></a>}}
\if{latex}{\out{\hypertarget{method-aggregated_voxels}{}}}
\subsection{Method \code{aggregated_voxels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VoxelDB$aggregated_voxels(
  ext = NULL,
  x = NULL,
  y = NULL,
  z = NULL,
  time_slice_id = NULL,
  time_slice_name = NULL
)}\if{html}{\out{</div>}}
}

}
}
